+++ previous-versions/assess/RELEASES.md
@@ -597,3 +597,13 @@ Based on:
 - [typescript v0.32.1] previous-versions/assess
 ### Releases
 - [NPM v0.32.1] https://www.npmjs.com/package/@codat/assess/v/0.32.1 - previous-versions/assess
+
+## 2023-09-05 00:19:36
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.33.0] previous-versions/assess
+### Releases
+- [NPM v0.33.0] https://www.npmjs.com/package/@codat/assess/v/0.33.0 - previous-versions/assess